+++ packages/import/server/Import.js
@@ -3,14 +3,9 @@ Kepler.extend({
 	Import: {
 		geojsonToPlace: function(feature, importName) {
 
-			var coords = feature.geometry.coordinates,
+			var geom = feature.geometry,
 				props = feature.properties,
-				loc = [];
-
-			if(feature.geometry.type==='Point') {
-				loc = [coords[1], coords[0]];
-			}
-			//TODO else centroid
+				loc = K.Util.geo.centroid(geom);
 
 			if(K.Util.valid.loc(loc)) {
 
@@ -18,6 +13,8 @@ Kepler.extend({
 					name: K.Util.sanitize.name(props.name || ''),
 					url: K.Util.sanitize.url(props.url || ''),
 					loc: K.Util.geo.locRound(loc, 8),
+					geometry: geom,
+					//TODO simplify
 					import: {
 						name: importName,
 						data: feature